feat: implement authentication controller with user login, registration, and password recovery; refactor user controller to remove authentication logic
